ticket/requirements.txt
langchain>=0.0.350
langchain-google-genai>=0.0.4
google-generativeai>=0.3.0
python-dotenv>=1.0.0
openai>=1.10.0
langchain-openai>=0.0.2
requests>=2.31.0
